+++ banano/__init__.py
@@ -42,4 +42,4 @@ __all__ = [
     "DependencyError",
     "ModelError",
 ]
-__version__ = "2.0.0"
+__version__ = "2.1.0"